MOCA/backend/requirements.txt
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
psycopg2-binary==2.9.3
python-dotenv==0.19.2
